splitting main and plot routine
