+++ Makefile
@@ -5,6 +5,7 @@ install:
 	cp cfg/* $(rootsys)/etc/root/plugins/FuriousPlotter/
 	cp macros/* $(rootsys)/share/root/macros/
 	cp draw.py $(rootsys)/bin/
+	cp plot_manager.py $(rootsys)/bin/
 	cp tree_manager.py $(rootsys)/bin/
 	cp fp_utils.py $(rootsys)/lib/root
 	cp operations.py $(rootsys)/lib/root
@@ -18,5 +19,6 @@ uninstall:
 	rm -r $(rootsys)/lib/root/fp_utils.py 
 	rm -r $(rootsys)/lib/root/operations.py
 	rm -r $(rootsys)/bin/draw.py
+	rm -r $(rootsys)/bin/plot_manager.py
 	rm -r $(rootsys)/bin/tree_manager.py
