+++ tree_manager.py
@@ -7,7 +7,7 @@ import argparse
 import os
 import subprocess
 import ROOT
-from  fp_utils import *
+from fp_utils import *
 
 ###---TTree manager class---------------------------------------------
 class FPTreeCreator:
@@ -148,55 +148,4 @@ class FPTreeCreator:
             ROOT.gROOT.ProcessLine(cname+'* '+key.replace(".", "_")+' = new '+cname+'('+tname+');')
         else:
             ROOT.gROOT.ProcessLine(cname+'* '+key.replace(".", "_")+' = new '+cname+'("'+tname+'","'+tname+'");')
-### MAIN ###
-if __name__ == "__main__":
-
-    ROOT.gROOT.SetBatch(True)
-    ROOT.gROOT.Reset()
-    if ROOT.gSystem.Load("CfgManagerDict.so") == -1:
-        ROOT.gSystem.Load("CfgManager/lib/CfgManagerDict.so")
-    
-    parser = argparse.ArgumentParser (description = 'Draw plots from ROOT files')
-    parser.add_argument('-m', '--mod', type=str, default='', help='config file modifiers')
-    parser.add_argument('-c', '--cfg', default='', help='cfg file')
-    parser.add_argument('--debug', action='store_true', help='print debug information')
-    
-    cmd_opts = parser.parse_args()
-
-    cfg = ROOT.CfgManager()
-    if cmd_opts.cfg != "":
-        cfg.ParseConfigFile(cmd_opts.cfg)
-    if cmd_opts.mod != "":
-        for config in cmd_opts.mod.split(','):
-            print(config)
-            cfg.ParseConfigString(config)
-
-    if cmd_opts.debug:
-        cfg.Print()    
-        
-    #---Load py/C++ plugins(for preproc, style and postproc)
-    plugin_funcs = {}
-    plugins = {"py" : ['operations'], "C" : [], "so" : [], "line" : []}    
-    if cfg.OptExist("draw.plugins"):        
-        for plugin in cfg.GetOpt(vstring)("draw.plugins"):
-            if ".py" == plugin[-3:]:
-                plugins["py"].append(plugin[:-3])
-            elif ".C" == plugin[-2:]:
-                plugins["C"].append(plugin)
-            elif ".so" == plugin[-3:]:
-                plugins["so"].append(plugin)
-            else:
-                plugins["line"].append(plugin)
-    for plugin in plugins["py"]:
-        plugin_module = __import__(plugin)        
-        for key, func in getattr(plugin_module, 'dictionary').items():
-            plugin_funcs[key] = func
-    for macro in plugins["C"]:
-        ROOT.gROOT.LoadMacro(macro) 
-    for lib in plugins["so"]:
-        ROOT.gSystem.Load(lib) 
-    processLines(plugins["line"])
-    
-    #---Make plots with FPPlots
-    for plot_name in cfg.GetOpt(vstring)("draw.trees"):
-        FPTreeCreator(cfg, plot_name, plugin_funcs)
+
